+++ projects/plugins/backup/src/js/components/Admin.js
@@ -1,11 +1,11 @@
 /**
  * External dependencies
  */
-import { Fragment, useState, useEffect } from '@wordpress/element';
+import { useState, useEffect } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import apiFetch from '@wordpress/api-fetch';
 import { useSelect } from '@wordpress/data';
-import { JetpackFooter, JetpackLogo } from '@automattic/jetpack-components';
+import { JetpackFooter, JetpackLogo, getRedirectUrl } from '@automattic/jetpack-components';
 
 /**
  * Internal dependencies
@@ -18,7 +18,7 @@ import { STORE_ID } from '../store';
 
 /* eslint react/react-in-jsx-scope: 0 */
 const Admin = () => {
-	const [ connectionStatus, renderConnectScreen ] = useConnection();
+	const [ connectionStatus, renderConnectScreen, renderConnectionStatusCard ] = useConnection();
 	const [ capabilities, setCapabilities ] = useState( null );
 	const [ capabilitiesError, setCapabilitiesError ] = useState( null );
 	const [ connectionLoaded, setConnectionLoaded ] = useState( false );
@@ -70,7 +70,7 @@ const Admin = () => {
 						</p>
 						<a
 							class="button"
-							href={ `https://wordpress.com/plans/${ domain }` }
+							href={ getRedirectUrl( 'backup-plugin-upgrade' ) }
 							target="_blank"
 							rel="noreferrer"
 						>
@@ -161,15 +161,7 @@ const Admin = () => {
 	};
 
 	const renderManageConnection = () => {
-		// TODO: Integrate connection management from Connection Package
-		return (
-			<Fragment>
-				<h2>{ __( 'Manage your connection', 'jetpack-backup' ) }</h2>
-				<p className="notice notice-success">
-					{ __( 'Site and User Connected.', 'jetpack-backup' ) }
-				</p>
-			</Fragment>
-		);
+		return renderConnectionStatusCard();
 	};
 
 	// Renders additional segments under the jp-hero area condition on having a backup plan
@@ -187,7 +179,7 @@ const Admin = () => {
 					{ ! capabilities.includes( 'backup-realtime' ) && (
 						<a
 							class="jp-cut"
-							href={ 'https://wordpress.com/checkout/' + domain + '/jetpack_backup_realtime' }
+							href={ getRedirectUrl( 'backup-plugin-realtime-upgrade', { site: domain } ) }
 						>
 							<span>
 								{ __(
@@ -210,7 +202,7 @@ const Admin = () => {
 					</p>
 					<p>
 						<a
-							href={ 'https://cloud.jetpack.com/activity-log/' + domain }
+							href={ getRedirectUrl( 'backup-plugin-activity-log', { site: domain } ) }
 							target="_blank"
 							rel="noreferrer"
 						>