+++ projects/plugins/jetpack/modules/carousel/jetpack-carousel.js
@@ -35,27 +35,47 @@
 			} );
 		}
 
-		function getBackgroundImage( imgEl ) {
+		function getAverageColor( imgEl ) {
 			var canvas = document.createElement( 'canvas' ),
-				context = canvas.getContext && canvas.getContext( '2d' );
+				context = canvas.getContext && canvas.getContext( '2d' ),
+				imgData,
+				width,
+				height,
+				length,
+				rgb = { r: 0, g: 0, b: 0 },
+				count = 0;
 
 			if ( ! imgEl ) {
-				return;
+				return rgb;
 			}
 
-			context.filter = 'blur(20px) ';
-			context.drawImage( imgEl, 0, 0 );
-			var url = canvas.toDataURL( 'image/png' );
-			canvas = null;
+			height = canvas.height = imgEl.naturalHeight || imgEl.offsetHeight || imgEl.height;
+			width = canvas.width = imgEl.naturalWidth || imgEl.offsetWidth || imgEl.width;
 
-			return url;
+			context.drawImage( imgEl, 0, 0 );
+			imgData = context.getImageData( 0, 0, width, height );
+
+			length = imgData.data.length;
+
+			for ( var i = 0; i < length; i += 4 ) {
+				rgb.r += imgData.data[ i ];
+				rgb.g += imgData.data[ i + 1 ];
+				rgb.b += imgData.data[ i + 2 ];
+				count++;
+			}
+
+			rgb.r = Math.floor( rgb.r / count );
+			rgb.g = Math.floor( rgb.g / count );
+			rgb.b = Math.floor( rgb.b / count );
+
+			return rgb;
 		}
 
 		return {
 			noop: noop,
 			texturize: texturize,
 			applyReplacements: applyReplacements,
-			getBackgroundImage: getBackgroundImage,
+			getAverageColor: getAverageColor,
 		};
 	} )();
 
@@ -769,11 +789,8 @@
 
 			loadFullImage( carousel.slides[ index ] );
 
-			if (
-				Number( jetpackCarouselStrings.display_background_image ) === 1 &&
-				! carousel.slides[ index ].backgroundImage
-			) {
-				loadBackgroundImage( carousel.slides[ index ] );
+			if ( Number( jetpackCarouselStrings.display_slide_background ) === 1 ) {
+				loadSlideBackgrounds();
 			}
 
 			domUtil.hide( carousel.caption );
@@ -1031,7 +1048,7 @@
 				}
 
 				if ( title ) {
-					var plainTitle = domUtil.convertToPlainText( title );
+					var plainTitle = domUtil.stripHTML( title );
 					titleElement.innerHTML = plainTitle;
 
 					if ( ! caption ) {
@@ -1256,31 +1273,59 @@
 			}
 		}
 
-		function loadBackgroundImage( slide ) {
-			var currentSlide = slide.el;
+		function loadSlideBackgrounds() {
+			applySlideBackground( '.swiper-slide-active' );
 
-			if ( swiper && swiper.slides ) {
-				currentSlide = swiper.slides[ swiper.activeIndex ];
+			setTimeout( function () {
+				applySlideBackground( '.swiper-slide-prev' );
+				applySlideBackground( '.swiper-slide-next' );
+			}, 200 );
+		}
+
+		function applySlideBackground( slideClass ) {
+			var slideEl = carousel.container.querySelector( slideClass );
+
+			if ( ! slideEl ) {
+				return;
 			}
 
-			var image = slide.attrs.originalElement;
-			var isLoaded = image.complete && image.naturalHeight !== 0;
+			// We're done if there's already a background image set.
+			if ( slideEl.style.backgroundImage ) {
+				return;
+			}
 
+			var image = slideEl.querySelector( 'img' );
+			if ( ! image ) {
+				return;
+			}
+
+			var isLoaded = image.complete && image.naturalHeight !== 0;
 			if ( isLoaded ) {
-				applyBackgroundImage( slide, currentSlide, image );
+				calculateSlideBackgroundCss( slideEl, image );
 				return;
 			}
 
 			image.onload = function () {
-				applyBackgroundImage( slide, currentSlide, image );
+				calculateSlideBackgroundCss( slideEl, image );
 			};
 		}
 
-		function applyBackgroundImage( slide, currentSlide, image ) {
-			var url = util.getBackgroundImage( image );
-			slide.backgroundImage = url;
-			currentSlide.style.backgroundImage = 'url(' + url + ')';
-			currentSlide.style.backgroundSize = 'cover';
+		function calculateSlideBackgroundCss( slideEl, image ) {
+			var rgb = util.getAverageColor( image );
+			slideEl.style.backgroundImage =
+				'linear-gradient( to bottom, rgba(' +
+				rgb.r +
+				',' +
+				rgb.g +
+				',' +
+				rgb.b +
+				', 0.5 ), rgba(' +
+				rgb.r +
+				', ' +
+				rgb.g +
+				', ' +
+				rgb.b +
+				', 0.25 ';
 		}
 
 		function clearCommentTextAreaValue() {
@@ -1446,6 +1491,7 @@
 			};
 
 			var data = domUtil.getJSONAttribute( gallery, 'data-carousel-extra' );
+			var tapTimeout;
 
 			if ( ! data ) {
 				return; // don't run if the default gallery functions weren't used
@@ -1512,7 +1558,7 @@
 				threshold: 5,
 			} );
 
-			swiper.on( 'slideChange', function () {
+			swiper.on( 'slideChange', function ( swiper ) {
 				var index;
 				// Swiper indexes slides from 1, plus when looping to left last slide ends up
 				// as 0 and looping to right first slide as total slides + 1. These are adjusted
@@ -1525,6 +1571,36 @@
 					index = swiper.activeIndex - 1;
 				}
 				selectSlideAtIndex( index );
+
+				carousel.overlay.classList.remove( 'jp-carousel-hide-controls' );
+			} );
+
+			swiper.on( 'zoomChange', function ( swiper, scale ) {
+				if ( scale > 1 ) {
+					carousel.overlay.classList.add( 'jp-carousel-hide-controls' );
+				}
+
+				if ( scale === 1 ) {
+					carousel.overlay.classList.remove( 'jp-carousel-hide-controls' );
+				}
+			} );
+
+			swiper.on( 'doubleTap', function ( swiper ) {
+				clearTimeout( tapTimeout );
+				if ( swiper.zoom.scale === 1 ) {
+					var zoomTimeout = setTimeout( function () {
+						carousel.overlay.classList.remove( 'jp-carousel-hide-controls' );
+						clearTimeout( zoomTimeout );
+					}, 150 );
+				}
+			} );
+
+			swiper.on( 'tap', function () {
+				if ( swiper.zoom.scale > 1 ) {
+					tapTimeout = setTimeout( function () {
+						carousel.overlay.classList.toggle( 'jp-carousel-hide-controls' );
+					}, 150 );
+				}
 			} );
 
 			domUtil.fadeIn( carousel.overlay, function () {