+++ projects/plugins/jetpack/modules/stats.php
@@ -65,8 +65,6 @@ function stats_load() {
 	add_action( 'manage_posts_custom_column', 'jetpack_stats_post_table_cell', 10, 2 );
 	add_action( 'manage_pages_custom_column', 'jetpack_stats_post_table_cell', 10, 2 );
 
-	require_once __DIR__ . '/stats/class-jetpack-stats-upgrade-nudges.php';
-	Jetpack_Stats_Upgrade_Nudges::init();
 }
 
 /**
@@ -467,6 +465,9 @@ function stats_admin_path() {
  * @return void
  */
 function stats_reports_load() {
+	require_once __DIR__ . '/stats/class-jetpack-stats-upgrade-nudges.php';
+	Jetpack_Stats_Upgrade_Nudges::init();
+
 	wp_enqueue_script( 'jquery' );
 	wp_enqueue_script( 'postbox' );
 	wp_enqueue_script( 'underscore' );